minor code refactor / tidy up
