--- sip/execution_framework_interface/docker_compose_generator/tests/workflow_stage_config.json
@@ -1,41 +1,0 @@
-{
-    "workflow":
-    [
-        {
-            "type": "",
-            "status": "",
-            "resources_required": {},
-            "resources_assinged": {},
-            "ee_config": {},
-            "app_config": {}
-        },
-
-        {
-            "type": "vis_send",
-            "status": "",
-            "resources_required": {
-                "compute": {
-                    "nodes": 2
-                },
-                "storage": {},
-                "network": {}
-            },
-            "resources_assigned": {},
-            "ee_config": {
-                "type": "docker_swarm",
-                "compose_template": "docker-compose.send.j2.yml",
-                "num_senders": 2
-            },
-            "app_config": {
-                "command_args": {
-                    "template": "send.j2.json"
-                },
-                "destination_hosts": ["ip1", "ip2"],
-                "destination_port": {
-                    "start": 41000
-                }
-            }
-        }
-
-    ]
-}
